adding requestBackend to handle all calls to DB backend
