+++ lib/findTicket.js
@@ -1,8 +1,4 @@
-import fetch from 'node-fetch'
-import { parseString } from 'xml2js'
-
-const endpoint = 'https://fahrkarten.bahn.de/mobile/dbc/xs.go?'
-
+import { requestBackend } from './requestBackend.js'
 /**
  * findTicket function
  * @param   string | DB ticket number (Auftragsnummer 12 digits)
@@ -15,50 +11,12 @@ export const findTicket = async (ticketNumber, lastName) => {
   const reqBody = '<?xml version="1.0"?><rqfindorder version="1.0">'
     + '<rqheader l="de" v="23080000" d="iPhone16.4.1" os="iOS_15.7.5" app="NAVIGATOR"/>'
     + '<rqorder on="' + ticketNumber + '"/><authname tln="' + lastName + '"/></rqfindorder>'
-  return fetch(endpoint, {
-    method: 'POST',
-    headers: {
-      'User-Agent': 'github.com/envake/db-tickets',
-      'Accept-Language': 'de-DE,de;q=0.9'
-    },
-    body: reqBody
-  }).then(response => {
-    if (!response.ok) {
-      throw new HTTPResponseError(response)
-    }
-    return response.text()
-  }).then(body => {
-    return new Promise((resolve, reject) => parseString(body, (err, result) => {
-      if (err) {
-        reject(err)
-      }
-      else {
-        resolve(result)
-      }
-    }))
-  }).then(json => {
-    if (json['rperror']) {
-      throw new XMLRPCError(json['rperror']['error'])
-    }
-    const data = json.rporderheadlist
+    return requestBackend(reqBody).then(json => {
+      const data = json.rporderheadlist
   
     const _order = data['orderheadlist'][0]['orderhead'][0]['$']
-    return _order
-  })
+    return _order})
 }
-class HTTPResponseError extends Error {
-    constructor(response, ...args) {
-      super(`HTTP Error Response: ${response.status} ${response.statusText}`, ...args)
-      this.response = response
-    }
-  }
   
-  class XMLRPCError extends Error {
-    constructor(response, ...args) {
-      super(`XML Error Response: 
-        ${response[0]['$']['nr']}
-        ${response[0]['txt'][0]}`, ...args)
-      this.response = response
-    }
-  }
+    
 export default findTicket
