+++ lib/queryTicket.js
@@ -1,8 +1,5 @@
-import fetch from 'node-fetch'
-import { parseString } from 'xml2js'
 import { parseLegs, price } from './parse.js'
-
-const endpoint = 'https://fahrkarten.bahn.de/mobile/dbc/xs.go?'
+import requestBackend from './requestBackend.js';
 
 /**
  * queryTicket function
@@ -23,35 +20,9 @@ const queryTicket = async (ticketNumber, lastName,kwid=null) => {
     }
      reqBody +='<authname tln="' + lastName + '"/></rqorderdetails>';
 
-  return fetch(endpoint, {
-    method: 'POST',
-    headers: {
-      'User-Agent': 'github.com/envake/db-tickets',
-      'Accept-Language': 'de-DE,de;q=0.9'
-    },
-    body: reqBody
-  })
-  .then(response => {
-    if (!response.ok) {
-      throw new HTTPResponseError(response)
-    }
-    return response.text()
-  })
-  .then(body => {
-    return new Promise((resolve, reject) => parseString(body, (err, result) => {
-      if (err) {
-        reject(err)
-      }
-      else {
-        resolve(result)
-      }
-    }))
-  })
+  return requestBackend(reqBody)
   .then(json => {
-    if (json['rperror']) {
-      throw new XMLRPCError(json['rperror']['error'])
-    }
-
+  
     const data = json.rporderdetails
